Debugged all of the scripts; should be ready to go now.

+++ create_initial_labeled_hatids.py
@@ -1,8 +1,10 @@
 import os, sys
-from utils import *
-from featureutils import *
+from utils.miscutils import *
+from utils.featureutils import *
 from settings import *
 import cPickle as pickle
+
+#overwrite = True
 
 # Make the directory for this model if it doesn't already exist
 if not os.path.exists(model_output_dir): os.makedirs(model_output_dir)
@@ -15,13 +17,15 @@ if os.path.exists(fname):
 	print "%s already exists. You should delete %s if you're sure you want to overwrite it."%(fname, fname)
 	sys.exit()
 
-gcvs_crossmatches 					= np.loadtxt(full_gcvs_match_cat_name, dtype=match_cat_dt)
+gcvs_crossmatches 					= np.loadtxt(full_gcvs_match_cat_fname , dtype=match_cat_dt)
 ids_to_fetch 						= GetHATIDsToFetch(gcvs_crossmatches['id'])
 
 categories 							= GetCategoriesForEachHATID(gcvs_crossmatches)
-good_gcvs_hatids_fname = "%s/rrlyr_classification/good_gcvs_hatids.list"%(parent_dir)
+good_gcvs_hatids_fname = "%s/good_gcvs_hatids.list"%(parent_dir)
 print "Pruning out bad ID's"
 if overwrite or not os.path.exists(good_gcvs_hatids_fname):
+	#print gcvs_crossmatches['id']
+	#print categories
 	hatids 								= GetGoodHATIDs(gcvs_crossmatches['id'], categories)
 	pickle.dump(hatids, open(good_gcvs_hatids_fname, 'wb'))
 else:
@@ -36,6 +40,7 @@ for vclass in vartypes_to_classify:
 	print vclass, sum([ 1 for ID in hatids if categories[ID] == vclass ])
 
 # Write initial labeled ids!
+if os.path.exists(fname): raise Exception("File %s already exists; delete first and then try again."%(fname))
 f = open(fname, 'w')
 for ID in hatids:
 	f.write("%-20s%-10i%-20s\n"%(ID, 0, categories[ID]))
